implement authorization using policy

+++ app/Http/Controllers/QuestionsController.php
@@ -9,6 +9,11 @@ use Illuminate\Http\Request;
 
 class QuestionsController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('auth',['except' => ['index','show']]);
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -67,6 +72,7 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
+        $this->authorize('update',$question);
         return view('questions.edit',compact('question'));
     }
 
@@ -79,6 +85,8 @@ class QuestionsController extends Controller
      */
     public function update(AskQuesitonRequest $request, Question $question)
     {
+        $this->authorize('update',$question);
+
         $question->update($request->only('title','body'));
         return redirect('/questions')->with('success','Your question has been updated.');
     }
@@ -91,8 +99,9 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
-        $question->delete();
+        $this->authorize('delete',$question);
 
+        $question->delete();
         return redirect('/questions')->with('success','Your questions has been deleted.');
     }
 }
